Add function to extract syllable count from a text
When the system's response is generated, a transcript is available,
allowing for faster syllable counting than from the wav file and in
parallel with the synthesis. That is why a second function to extract
syllables was introduced. The current solution uses a 3rd party
implementation of Frank Liang's algorithm for hyphenation based on the
assumption that syllables are equal to the units that can be separated
by hyphens (which should be roughly correct).

.gitignore was also adjusted to ignore __pycache__ folders anywhere