Replace implementation of extract_syllables_text
The hyphenation algorithm previously used to determine the number of
syllables in a text did not work well for proper nouns and many other
cases. This new implementation is much better and even outperforms
AuToBI (might be removed as well later).
Also moved file name suffix code to auxiliary function.